Upgrade to React 16 and move to TSX
