+++ circle.yml
@@ -1,15 +1,6 @@
 machine:
   node:
     version: 6.5.0
-deployment:
-  npm-publish:
-    tag: /v[0-9]+(\.[0-9]+){2}(-rc[0-9]+)?/
-    owner: palantir
-    commands:
-      - touch .npmrc
-      - printf "\n$NPM_REGISTRY_CREDS\n" >> .npmrc
-      - yarn run build
-      - npm publish
 dependencies:
   pre:
     - npm install -g yarn
@@ -22,11 +13,15 @@ dependencies:
 test:
   override:
     - yarn run build
-    - "yarn run test -- --reporter mocha-junit-reporter":
+    - yarn run test:
         environment:
           MOCHA_FILE: $CIRCLE_TEST_REPORTS/junit/test-results.xml
-general:
-  branches:
-    ignore:
-      - gh-pages # list of branches to ignore
-
+deployment:
+  npm-publish:
+    tag: /v[0-9]+(\.[0-9]+){2}(-rc[0-9]+)?/
+    owner: palantir
+    commands:
+      - touch .npmrc
+      - printf "\n$NPM_REGISTRY_CREDS\n" >> .npmrc
+      - yarn run build
+      - npm publish
