+++ dev/index.tsx
@@ -14,7 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { ExampleApp } from './ExampleApp';
 
-ReactDOM.render(ExampleApp(), document.getElementById('app')!);
+ReactDOM.render(<ExampleApp/>, document.getElementById('app')!);
