+++ package.json
@@ -13,65 +13,69 @@
   "scripts": {
     "clean": "rm -rf lib/",
     "start": "webpack-dev-server --config webpack.config.hot.js --inline --hot --content-base docs/ --host 0.0.0.0 --port 8092",
-    "build": "npm-run-all clean lib.compile.ts lib.compile.less lib.remove.type.references",
+    "build": "npm-run-all clean build:ts build:less",
+    "build:ts": "tsc -p tsconfig-build.json",
+    "build:less": " lessc --source-map styles/index.less react-mosaic-component.css",
     "docs": "rm -rf docs/ && webpack",
-    "test": "tslint --project tsconfig.json && mocha --opts test/mocha.opts 'test/*.ts'",
-    "lib.compile.ts": "tsc -p tsconfig-build.json",
-    "lib.compile.less": " lessc --source-map styles/index.less react-mosaic-component.css",
-    "lib.remove.type.references": "replace '^/// <reference.+?\\n' '' lib -r --include='*.d.ts'"
+    "test": "npm-run-all test:lint test:unit",
+    "test:circle": "npm-run-all test:lint test:unit:ci",
+    "test:lint": "tslint --project tsconfig.json",
+    "test:unit": "mocha --opts test/mocha.opts 'test/*.ts'",
+    "test:unit:ci": "mocha --opts test/mocha.opts --reporter mocha-junit-reporter 'test/*.ts'"
   },
   "devDependencies": {
-    "@blueprintjs/core": "^1.23.1",
+    "@blueprintjs/core": "^1.29.0",
     "@types/assertion-error": "^1.0.30",
-    "@types/chai": "^3.4.34",
-    "@types/classnames": "^0.0.32",
-    "@types/es6-promise": "^0.0.32",
+    "@types/chai": "^4.0.4",
+    "@types/classnames": "^2.2.3",
+    "@types/es6-promise": "^0.0.33",
     "@types/immutability-helper": "^2.0.15",
-    "@types/lodash": "^4.14.43",
-    "@types/mocha": "^2.2.33",
-    "@types/prop-types": "^15.5.1",
-    "@types/react": "^15.0.38",
-    "@types/react-dnd": "^2.0.33",
+    "@types/lodash": "^4.14.76",
+    "@types/mocha": "^2.2.43",
+    "@types/prop-types": "^15.5.2",
+    "@types/react": "^16.0.7",
+    "@types/react-dnd": "^2.0.34",
     "@types/react-dnd-html5-backend": "^2.1.8",
-    "@types/react-dom": "^15.5.1",
-    "@types/uuid": "^2.0.29",
-    "chai": "^3.5.0",
-    "css-loader": "^0.26.1",
-    "file-loader": "^0.9.0",
-    "html-loader": "^0.4.4",
-    "html-webpack-plugin": "^2.24.1",
-    "jsdom": "^9.8.3",
-    "jsdom-global": "^2.1.0",
+    "@types/react-dom": "^15.5.4",
+    "@types/uuid": "^3.4.2",
+    "chai": "^4.1.2",
+    "css-loader": "^0.28.7",
+    "file-loader": "^0.11.2",
+    "html-loader": "^0.5.1",
+    "html-webpack-plugin": "^2.30.1",
+    "jsdom": "^11.2.0",
+    "jsdom-global": "^3.0.2",
     "less": "^2.6.0",
-    "less-loader": "^2.2.2",
-    "mocha": "^3.2.0",
+    "less-loader": "^2.2.3",
+    "mocha": "^3.5.3",
     "mocha-junit-reporter": "^1.12.1",
-    "npm-run-all": "^4.0.0",
-    "react": "^15.6.1",
-    "react-addons-css-transition-group": "^15.6.0",
-    "react-dom": "^15.6.1",
+    "npm-run-all": "^4.1.1",
+    "react": "^16.0.0",
+    "react-addons-css-transition-group": "^15.6.2",
+    "react-dom": "^16.0.0",
     "react-hot-loader": "^1.3.1",
     "replace": "v0.3.0",
-    "source-map-loader": "^0.1.5",
-    "style-loader": "^0.13.0",
-    "ts-loader": "^2.3.1",
-    "ts-node": "^3.2.0",
-    "tslint": "^5.0.0",
-    "typescript": "^2.4.1",
+    "source-map-loader": "^0.2.1",
+    "style-loader": "^0.18.2",
+    "ts-loader": "^2.3.7",
+    "ts-node": "^3.3.0",
+    "tslint": "^5.7.0",
+    "typescript": "^2.5.2",
     "webpack": "^1.14.0",
     "webpack-dev-server": "^1.16.2"
   },
   "dependencies": {
+    "@types/react-hot-loader": "^3.0.4",
     "classnames": "^2.2.5",
-    "immutability-helper": "^2.0.0",
+    "immutability-helper": "^2.4.0",
     "lodash": "^4.17.2",
-    "prop-types": "^15.5.10",
-    "react-dnd": "^2.1.4",
-    "react-dnd-html5-backend": "^2.1.2",
+    "prop-types": "^15.6.0",
+    "react-dnd": "^2.5.1",
+    "react-dnd-html5-backend": "^2.5.1",
     "uuid": "^3.0.1"
   },
   "peerDependencies": {
-    "react": "^15.0.0 || ^0.14.0"
+    "react": "^15.0.0 || ^16.0.0"
   },
   "keywords": [
     "ui",
