+++ src/index.ts
@@ -15,69 +15,69 @@
  * limitations under the License.
  */
 export {
-    Mosaic,
-    MosaicProps,
-    MosaicUncontrolledProps,
-    MosaicControlledProps,
-    MosaicFactory,
-    MosaicWithoutDragDropContext,
+  Mosaic,
+  MosaicProps,
+  MosaicUncontrolledProps,
+  MosaicControlledProps,
+  MosaicFactory,
+  MosaicWithoutDragDropContext,
 } from './Mosaic';
 export {
-    MosaicNode,
-    MosaicDragType,
-    MosaicDirection,
-    MosaicBranch,
-    CreateNode,
-    MosaicParent,
-    MosaicPath,
-    MosaicUpdate,
-    MosaicUpdateSpec,
-    TileRenderer,
+  MosaicNode,
+  MosaicDragType,
+  MosaicDirection,
+  MosaicBranch,
+  CreateNode,
+  MosaicParent,
+  MosaicPath,
+  MosaicUpdate,
+  MosaicUpdateSpec,
+  TileRenderer,
 } from './types';
 export {
-    MosaicContext,
-    MosaicActionsPropType,
-    MosaicPathGetterPropType,
-    MosaicRootActions,
-    MosaicTileContext,
-    MosaicWindowActions,
-    MosaicWindowActionsPropType,
-    MosaicWindowContext,
+  MosaicContext,
+  MosaicActionsPropType,
+  MosaicPathGetterPropType,
+  MosaicRootActions,
+  MosaicTileContext,
+  MosaicWindowActions,
+  MosaicWindowActionsPropType,
+  MosaicWindowContext,
 } from './contextTypes';
 export {
-    buildSpecFromUpdate,
-    createDragToUpdates,
-    createExpandUpdate,
-    createHideUpdate,
-    createRemoveUpdate,
-    updateTree,
+  buildSpecFromUpdate,
+  createDragToUpdates,
+  createExpandUpdate,
+  createHideUpdate,
+  createRemoveUpdate,
+  updateTree,
 } from './mosaicUpdates';
 export {
-    createBalancedTreeFromLeaves,
-    Corner,
-    getAndAssertNodeAtPathExists,
-    getLeaves,
-    getNodeAtPath,
-    getOtherBranch,
-    getOtherDirection,
-    getPathToCorner,
-    isParent,
+  createBalancedTreeFromLeaves,
+  Corner,
+  getAndAssertNodeAtPathExists,
+  getLeaves,
+  getNodeAtPath,
+  getOtherBranch,
+  getOtherDirection,
+  getPathToCorner,
+  isParent,
 } from './mosaicUtilities';
 export { MosaicWindow, MosaicWindowFactory, MosaicWindowProps } from './window/MosaicWindow';
 export {
-    createDefaultToolbarButton,
-    DEFAULT_CONTROLS_WITH_CREATION,
-    DEFAULT_CONTROLS_WITHOUT_CREATION,
-    ExpandButton,
-    ExpandButtonFactory,
-    MosaicButtonProps,
-    RemoveButton,
-    RemoveButtonFactory,
-    ReplaceButton,
-    ReplaceButtonFactory,
-    Separator,
-    SeparatorFactory,
-    SplitButton,
-    SplitButtonFactory,
+  createDefaultToolbarButton,
+  DEFAULT_CONTROLS_WITH_CREATION,
+  DEFAULT_CONTROLS_WITHOUT_CREATION,
+  ExpandButton,
+  ExpandButtonFactory,
+  MosaicButtonProps,
+  RemoveButton,
+  RemoveButtonFactory,
+  ReplaceButton,
+  ReplaceButtonFactory,
+  Separator,
+  SeparatorFactory,
+  SplitButton,
+  SplitButtonFactory,
 } from './window/defaultToolbarControls';
 export { MosaicZeroState, MosaicZeroStateFactory, MosaicZeroStateProps } from './MosaicZeroState';
