+++ src/mosaicUtilities.ts
@@ -18,23 +18,23 @@ import * as _ from 'lodash';
 import { MosaicBranch, MosaicDirection, MosaicNode, MosaicParent, MosaicPath } from './types';
 
 function alternateDirection<T>(node: MosaicNode<T>, direction: MosaicDirection = 'row'): MosaicNode<T> {
-    if (isParent(node)) {
-        const nextDirection = getOtherDirection(direction);
-        return {
-            direction,
-            first: alternateDirection(node.first, nextDirection),
-            second: alternateDirection(node.second, nextDirection),
-        };
-    } else {
-        return node;
-    }
+  if (isParent(node)) {
+    const nextDirection = getOtherDirection(direction);
+    return {
+      direction,
+      first: alternateDirection(node.first, nextDirection),
+      second: alternateDirection(node.second, nextDirection),
+    };
+  } else {
+    return node;
+  }
 }
 
 export enum Corner {
-    TOP_LEFT = 1,
-    TOP_RIGHT,
-    BOTTOM_LEFT,
-    BOTTOM_RIGHT,
+  TOP_LEFT = 1,
+  TOP_RIGHT,
+  BOTTOM_LEFT,
+  BOTTOM_RIGHT,
 }
 
 /**
@@ -43,7 +43,7 @@ export enum Corner {
  * @returns {boolean}
  */
 export function isParent<T>(node: MosaicNode<T>): node is MosaicParent<T> {
-    return (node as MosaicParent<T>).direction != null;
+  return (node as MosaicParent<T>).direction != null;
 }
 
 /**
@@ -54,29 +54,29 @@ export function isParent<T>(node: MosaicNode<T>): node is MosaicParent<T> {
  */
 export function createBalancedTreeFromLeaves<T>(leaves: MosaicNode<T>[],
                                                 startDirection: MosaicDirection = 'row'): MosaicNode<T> | null {
-    if (leaves.length === 0) {
-        return null;
-    }
+  if (leaves.length === 0) {
+    return null;
+  }
 
-    let current: MosaicNode<T>[] = _.clone(leaves);
-    let next: MosaicNode<T>[] = [];
+  let current: MosaicNode<T>[] = _.clone(leaves);
+  let next: MosaicNode<T>[] = [];
 
-    while (current.length > 1) {
-        while (current.length > 0) {
-            if (current.length > 1) {
-                next.push({
-                    direction: 'row',
-                    first: current.shift()!,
-                    second: current.shift()!,
-                });
-            } else {
-                next.unshift(current.shift()!);
-            }
-        }
-        current = next;
-        next = [];
+  while (current.length > 1) {
+    while (current.length > 0) {
+      if (current.length > 1) {
+        next.push({
+          direction: 'row',
+          first: current.shift()!,
+          second: current.shift()!,
+        });
+      } else {
+        next.unshift(current.shift()!);
+      }
     }
-    return alternateDirection(current[0], startDirection);
+    current = next;
+    next = [];
+  }
+  return alternateDirection(current[0], startDirection);
 }
 
 /**
@@ -85,13 +85,13 @@ export function createBalancedTreeFromLeaves<T>(leaves: MosaicNode<T>[],
  * @returns {any}
  */
 export function getOtherBranch(branch: MosaicBranch): MosaicBranch {
-    if (branch === 'first') {
-        return 'second';
-    } else if (branch === 'second') {
-        return 'first';
-    } else {
-        throw new Error(`Branch '${branch}' not a valid branch`);
-    }
+  if (branch === 'first') {
+    return 'second';
+  } else if (branch === 'second') {
+    return 'first';
+  } else {
+    throw new Error(`Branch '${branch}' not a valid branch`);
+  }
 }
 
 /**
@@ -100,11 +100,11 @@ export function getOtherBranch(branch: MosaicBranch): MosaicBranch {
  * @returns {any}
  */
 export function getOtherDirection(direction: MosaicDirection): MosaicDirection {
-    if (direction === 'row') {
-        return 'column';
-    } else {
-        return 'row';
-    }
+  if (direction === 'row') {
+    return 'column';
+  } else {
+    return 'row';
+  }
 }
 
 /**
@@ -114,22 +114,22 @@ export function getOtherDirection(direction: MosaicDirection): MosaicDirection {
  * @returns {MosaicPath}
  */
 export function getPathToCorner(tree: MosaicNode<any>, corner: Corner): MosaicPath {
-    let currentNode: MosaicNode<any> = tree;
-    const currentPath: MosaicPath = [];
-    while (isParent(currentNode)) {
-        if (currentNode.direction === 'row' && (corner === Corner.TOP_LEFT || corner === Corner.BOTTOM_LEFT)) {
-            currentPath.push('first');
-            currentNode = currentNode.first;
-        } else if (currentNode.direction === 'column' && (corner === Corner.TOP_LEFT || corner === Corner.TOP_RIGHT)) {
-            currentPath.push('first');
-            currentNode = currentNode.first;
-        } else {
-            currentPath.push('second');
-            currentNode = currentNode.second;
-        }
+  let currentNode: MosaicNode<any> = tree;
+  const currentPath: MosaicPath = [];
+  while (isParent(currentNode)) {
+    if (currentNode.direction === 'row' && (corner === Corner.TOP_LEFT || corner === Corner.BOTTOM_LEFT)) {
+      currentPath.push('first');
+      currentNode = currentNode.first;
+    } else if (currentNode.direction === 'column' && (corner === Corner.TOP_LEFT || corner === Corner.TOP_RIGHT)) {
+      currentPath.push('first');
+      currentNode = currentNode.first;
+    } else {
+      currentPath.push('second');
+      currentNode = currentNode.second;
     }
+  }
 
-    return currentPath;
+  return currentPath;
 }
 
 /**
@@ -138,14 +138,14 @@ export function getPathToCorner(tree: MosaicNode<any>, corner: Corner): MosaicPa
  * @returns {T[]}
  */
 export function getLeaves<T>(tree: MosaicNode<T> | null): T[] {
-    if (tree == null) {
-        return [];
-    } else if (isParent(tree)) {
-        return getLeaves(tree.first)
-            .concat(getLeaves(tree.second));
-    } else {
-        return [tree];
-    }
+  if (tree == null) {
+    return [];
+  } else if (isParent(tree)) {
+    return getLeaves(tree.first)
+      .concat(getLeaves(tree.second));
+  } else {
+    return [tree];
+  }
 }
 
 /**
@@ -155,11 +155,11 @@ export function getLeaves<T>(tree: MosaicNode<T> | null): T[] {
  * @returns {MosaicNode<T>|null}
  */
 export function getNodeAtPath<T>(tree: MosaicNode<T> | null, path: MosaicPath): MosaicNode<T> | null {
-    if (path.length > 0) {
-        return _.get<MosaicNode<T>>(tree, path, null!);
-    } else {
-        return tree;
-    }
+  if (path.length > 0) {
+    return _.get<MosaicNode<T>>(tree, path, null!);
+  } else {
+    return tree;
+  }
 }
 
 /**
@@ -169,12 +169,12 @@ export function getNodeAtPath<T>(tree: MosaicNode<T> | null, path: MosaicPath): 
  * @returns {MosaicNode<T>}
  */
 export function getAndAssertNodeAtPathExists<T>(tree: MosaicNode<T> | null, path: MosaicPath): MosaicNode<T> {
-    if (tree == null) {
-        throw new Error('Root is empty, cannot fetch path');
-    }
-    const node = getNodeAtPath(tree, path);
-    if (node == null) {
-        throw new Error(`Path [${path.join(', ')}] did not resolve to a node`);
-    }
-    return node;
+  if (tree == null) {
+    throw new Error('Root is empty, cannot fetch path');
+  }
+  const node = getNodeAtPath(tree, path);
+  if (node == null) {
+    throw new Error(`Path [${path.join(', ')}] did not resolve to a node`);
+  }
+  return node;
 }
