+++ tsconfig-build.json
@@ -1,15 +1,15 @@
 {
-    "compilerOptions": {
-        "rootDir": "src",
-        "declaration": true
-    },
-    "exclude": [
-        "dev",
-        "docs",
-        "lib",
-        "node_modules",
-        "test",
-        "webpack.config*.js"
-    ],
-    "extends": "./tsconfig.json"
+  "compilerOptions": {
+    "rootDir": "src",
+    "declaration": true
+  },
+  "exclude": [
+    "dev",
+    "docs",
+    "lib",
+    "node_modules",
+    "test",
+    "webpack.config*.js"
+  ],
+  "extends": "./tsconfig.json"
 }
